Restructure VID code for Python/FWLite
* Avoid building dictionaries for tiny helper classes that can also be
  defined with `ROOT.gInterpreter.Declare` on the fly.

* Define dictionaries for `VersionedSelector<>` template instantiations
  in PhysicsTools/SelectorUtils. Like this we don't need to generate
  dictionaries for ElectronIdentification, MuonIdentification, and
  PhotonIdentification, all the dictionaries for VersionedSelectors are
  in one place, and we reduce the dependency of Reco packages on
  PhysicsTools.
