+++ PhysicsTools/SelectorUtils/src/classes.h
@@ -10,14 +10,12 @@
 #include "PhysicsTools/SelectorUtils/interface/PFJetIDSelectionFunctor.h"
 #include "PhysicsTools/SelectorUtils/interface/PVSelector.h"
 #include "PhysicsTools/SelectorUtils/interface/RunLumiSelector.h"
-#include "PhysicsTools/SelectorUtils/interface/MakePyVIDClassBuilder.h"
 
-namespace PhysicsTools_SelectorUtils {
-  struct dictionary {
-    pat::strbitset strbitset;
-    edm::Wrapper<pat::strbitset> wstrbitset;
-    std::vector<pat::strbitset> vstrbitset;
-    edm::Wrapper<std::vector<pat::strbitset> > wvstrbitset;
-  };
-
-}  // namespace PhysicsTools_SelectorUtils
+#include "PhysicsTools/SelectorUtils/interface/VersionedSelector.h"
+#include "DataFormats/PatCandidates/interface/Photon.h"
+#include "DataFormats/EgammaCandidates/interface/PhotonFwd.h"
+#include "DataFormats/MuonReco/interface/MuonFwd.h"
+#include "DataFormats/PatCandidates/interface/Muon.h"
+#include "DataFormats/EgammaCandidates/interface/GsfElectronFwd.h"
+#include "DataFormats/EgammaCandidates/interface/GsfElectron.h"
+#include "DataFormats/PatCandidates/interface/Electron.h"
