--- RecoEgamma/ElectronIdentification/src/classes_def.xml
@@ -1,7 +1,0 @@
-<lcgdict>  
-  <class pattern="VersionedSelector<*>"/>
-  <class pattern="MakeVersionedSelector<*>"/>
-  <class pattern="PrintVIDToString<*>"/>
-  <class pattern="MakePtrFromCollection<*>"/>
-  <class name="ElectronMVAEstimatorRun2" transient="true"/>
-</lcgdict>
