+++ RecoEgamma/PhotonIdentification/python/VIDPhotonSelector.py
@@ -7,6 +7,5 @@ class VIDPhotonSelector(VIDSelectorBase):
         ptrmaker = ROOT.MakePtrFromCollection(ROOT.vector(ROOT.pat.Photon),
                                               ROOT.pat.Photon,
                                               ROOT.reco.Photon)
-        printer  = ROOT.PrintVIDToString(ROOT.reco.Photon)
-        VIDSelectorBase.__init__(self,builder,ptrmaker,printer,pythonpset)
+        VIDSelectorBase.__init__(self,builder,ptrmaker,pythonpset)
