+++ RecoMuon/MuonIdentification/python/VIDMuonSelector.py
@@ -7,6 +7,5 @@ class VIDMuonSelector(VIDSelectorBase):
         ptrmaker = ROOT.MakePtrFromCollection(ROOT.vector(ROOT.pat.Muon),
                                               ROOT.pat.Muon,
                                               ROOT.reco.Muon)
-        printer  = ROOT.PrintVIDToString(ROOT.reco.Muon)
-        VIDSelectorBase.__init__(self,builder,ptrmaker,printer,pythonpset)
+        VIDSelectorBase.__init__(self,builder,ptrmaker,pythonpset)
