--- RecoMuon/MuonIdentification/src/classes_def.xml
@@ -1,6 +1,0 @@
-<lcgdict>  
-  <class pattern="VersionedSelector<*>"/>
-  <class pattern="MakeVersionedSelector<*>"/>
-  <class pattern="PrintVIDToString<*>"/>
-  <class pattern="MakePtrFromCollection<*>"/>
-</lcgdict>
